move swipecontrols patchfiles to correct folders
